+++ sphinx/source/changelog.rst
@@ -16,6 +16,8 @@ Changelog
   a TAP/TZX file in a zip archive, in case there is more than one)
 * Added the ``--tape-start`` and ``--tape-stop`` options to :ref:`tap2sna.py`
   (to start or stop the tape at a specific block number)
+* Added the ``--tape-sum`` option to :ref:`tap2sna.py` (to specify the MD5
+  checksum of the TAP/TZX file)
 * Added the ``--interrupts`` option to :ref:`trace.py` (to enable the execution
   of interrupt routines)
 * :ref:`trace.py` now reads and writes the T-states counter in Z80 snapshots